requirements.txt
djangorestframework==3.10.3
Django==2.2.5
django-cors-middleware==1.4.0
Pygments==2.2.0
bcrypt==3.1.7
redgreenunittest==0.1.1
psycopg2-binary==2.8.3
gunicorn==19.9.0
sendgrid==6.1.0
drf-yasg==1.17.1
django-rest-passwordreset==1.1.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.1
django-countries==6.1.2
django-heroku==0.3.1
pyjwt==1.7.1
pydash==4.9.2
whitenoise==5.0.1
python-http-client==3.2.1
dj-database-url==0.5.0
django-filter==2.2.0
celery==4.4.0
django_celery_beat==1.4.0
kombu==4.6.7
more-itertools==8.1.0
pytz==2019.3
redis==3.3.11
amqp==2.5.2
redgreenunittest==0.1.1